refactored Availability section
